requirements.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click-didyoumean==0.0.3
click-repl==0.1.6
Django==2.2.11
dparse==0.5.1
feedparser==6.0.2
idna==2.10
importlib-metadata==3.1.0
kombu==5.0.2
packaging==20.4
prompt-toolkit==3.0.8
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
safety==1.9.0
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
django-extensions==3.1.0
graphene-django==2.13.0
django-filter==2.4.0
gevent==20.9.0
pytest-django==4.1.0
factory-boy==3.1.0
django-redis==4.12.1
pytest-factoryboy==2.0.3